# eskip: improve lexer performance 2